rename : #9 enums, embedded 디렉터리 추가
